requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.0
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
onnxruntime==1.18.1
# Opcional: para backend Ultralytics (PyTorch). Mantener solo si deseas usar MODEl_BACKEND=ultralytics
ultralytics>=8.2.100
requests==2.32.3
Pillow==10.4.0
transformers>=4.46.0
torch
torchvision
autoawq>=0.2.0
accelerate>=0.26.0
compressed-tensors
decord
sentencepiece